+++ index.js
@@ -12,6 +12,7 @@ const fs = require('fs')
  * Locals
  */
 
+let filename
 let now = new Date()
 let year = (now.getFullYear() < 10 ? '0' : '') + now.getFullYear()
 let month = ((now.getMonth() + 1) < 10 ? '0' : '') + (now.getMonth() + 1)
@@ -19,8 +20,17 @@ let day = (now.getDate() < 10 ? '0' : '') + now.getDate()
 let hours = (now.getHours() < 10 ? '0' : '') + now.getHours()
 let minutes = (now.getMinutes() < 10 ? '0' : '') + now.getMinutes()
 let seconds = (now.getSeconds() < 10 ? '0' : '') + now.getSeconds()
-
 let timestamp = `${year}${month}${day}${hours}${minutes}${seconds}`
+
+/**
+ * Accept filename as argument
+ */
+
+if (process.argv[2]) {
+  filename = `_${process.argv[2]}.js`
+} else {
+  filename = '.js'
+}
 
 /**
  * Ensure ./db directory exists
@@ -37,7 +47,7 @@ fs.stat('./db/migrations', (err, stats) => {
  * Create migration file
  */
 
-fs.writeFile(`./db/migrations/${timestamp}`, "'use strict'\n", {
+fs.writeFile(`./db/migrations/${timestamp}${filename}`, "'use strict'\n", {
   "encoding": "utf8",
   "flag": 'wx',
   "mode": 0o755